requirements.txt
absl-py==2.4.0
altair==6.0.0
anyio==4.12.1
asgiref==3.8.1
astroid==3.0.1
astunparse==1.6.3
attrs==25.4.0
autopep8==2.0.4
blinker==1.9.0
cachetools==7.0.5
certifi==2026.1.4
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.7
distlib==0.3.8
Django==5.0.7
filelock==3.15.4
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.53.0
fsspec==2026.1.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.7
idna==3.11
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiwer==4.0.0
joblib==1.4.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.5
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.18.0
networkx==3.6.1
numpy==2.0.0
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==24.1
pandas==2.2.2
pep8==1.7.1
pillow==10.3.0
platformdirs==3.11.0
protobuf==6.33.5
pyarrow==23.0.1
pycodestyle==2.11.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pylint==3.0.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.5.0
scipy==1.13.1
sentence-transformers==5.2.2
setuptools==80.10.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.3
sqlparse==0.5.0
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.5.0
tokenizers==0.22.2
toml==0.10.2
tomlkit==0.12.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.6.3
virtualenv==20.26.3
watchdog==6.0.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
